requirements.txt
pillow>=12.0.0
pytest>=8.0.0
openai>=1.0.0
